ProjectManagement/requirements.txt
ipykernel
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt